+++ src/pessoas/model/FactoryPessoa.java
@@ -7,13 +7,13 @@ public class FactoryPessoa {
 	public Pessoa setPessoa (String nomeCompleto, int dia, int mes, int ano, long cpf, 
 			String enderecoRua, String enderecoNumero, String enderecoComplemento, String enderecoBairro, 
 			String enderecoCep, String enderecoCidade, String enderecoEstado,  String enderecoPais,
-			int telefone, String tipo) {
+			long telefone, String tipo) {
 		return null;
 	}
 
 	public Cliente setPessoa(String nomeCompleto, int dia, int mes, int ano, long cpf, 
 			String enderecoRua, String enderecoNumero, String enderecoComplemento, String enderecoBairro, 
-			String enderecoCep, String enderecoCidade, String enderecoEstado,  String enderecoPais, int telefone, String tipo, boolean spc) {
+			String enderecoCep, String enderecoCidade, String enderecoEstado,  String enderecoPais, long telefone, String tipo, boolean spc) {
 		try {
 			return new Cliente(nomeCompleto, dia, mes, ano, cpf, 
 					enderecoRua, enderecoNumero, enderecoComplemento, enderecoBairro, 
@@ -27,7 +27,7 @@ public class FactoryPessoa {
 	public Funcionario setPessoa(String nomeCompleto, int dia, int mes, int ano, int cpf,
 			String enderecoRua, String enderecoNumero, String enderecoComplemento, String enderecoBairro, 
 			String enderecoCep, String enderecoCidade, String enderecoEstado,  String enderecoPais,
-			 int telefone, int diaContratacao, int mesContratacao, int anoContratacao, 
+			long telefone, int diaContratacao, int mesContratacao, int anoContratacao, 
 			 String senha, String login, boolean admin) {
 		try {
 			return new Funcionario(nomeCompleto, dia, mes, ano,  cpf,
